requirements/base.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
psycopg==3.1.9
channels==4.0.0
channels-redis==4.1.0
Pillow==10.0.0
celery==5.3.1
redis==5.0.0
django-treebeard==4.7
gunicorn